+++ crates/rust-client/src/sync/mod.rs
@@ -1,0 +1,4 @@
+mod sync;
+pub use sync::*;
+mod block_headers;
+mod tags;
